RDC-987: patching 7.1.4 -> 7.1.5

+++ src/main/java/net/ravendb/client/documents/operations/AI/AiUsage.java
@@ -4,16 +4,16 @@ import com.fasterxml.jackson.databind.JsonNode;
 import net.ravendb.client.Constants;
 
 public class AiUsage {
-    private int promptTokens;
-    private int completionTokens;
-    private int totalTokens;
-    private int cachedTokens;
-    private int reasoningTokens;
+    private long promptTokens;
+    private long completionTokens;
+    private long totalTokens;
+    private long cachedTokens;
+    private long reasoningTokens;
 
     public AiUsage() {
     }
 
-    public AiUsage(int promptTokens, int completionTokens, int totalTokens, int cachedTokens, int reasoningTokens) {
+    public AiUsage(long promptTokens, long completionTokens, long totalTokens, long cachedTokens, long reasoningTokens) {
         this.promptTokens = promptTokens;
         this.completionTokens = completionTokens;
         this.totalTokens = totalTokens;
@@ -21,66 +21,66 @@ public class AiUsage {
         this.reasoningTokens = reasoningTokens;
     }
 
-    public int getReasoningTokens() { return reasoningTokens; }
+    public long getReasoningTokens() { return reasoningTokens; }
 
-    public void setReasoningTokens(int reasoningTokens) { this.reasoningTokens = reasoningTokens; }
+    public void setReasoningTokens(long reasoningTokens) { this.reasoningTokens = reasoningTokens; }
 
-    public int getPromptTokens() {
+    public long getPromptTokens() {
         return promptTokens;
     }
 
-    public void setPromptTokens(int promptTokens) {
+    public void setPromptTokens(long promptTokens) {
         this.promptTokens = promptTokens;
     }
 
-    public int getCompletionTokens() {
+    public long getCompletionTokens() {
         return completionTokens;
     }
 
-    public void setCompletionTokens(int completionTokens) {
+    public void setCompletionTokens(long completionTokens) {
         this.completionTokens = completionTokens;
     }
 
-    public int getTotalTokens() {
+    public long getTotalTokens() {
         return totalTokens;
     }
 
-    public void setTotalTokens(int totalTokens) {
+    public void setTotalTokens(long totalTokens) {
         this.totalTokens = totalTokens;
     }
 
-    public int getCachedTokens() {
+    public long getCachedTokens() {
         return cachedTokens;
     }
 
-    public void setCachedTokens(int cachedTokens) {
+    public void setCachedTokens(long cachedTokens) {
         this.cachedTokens = cachedTokens;
     }
 
     void updateFrom(JsonNode json) {
         if (json.has(Constants.AI.PROMPT_TOKENS)) {
-            this.promptTokens += json.get(Constants.AI.PROMPT_TOKENS).asInt();
+            this.promptTokens += json.get(Constants.AI.PROMPT_TOKENS).asLong();
         }
         if (json.has(Constants.AI.COMPLETION_TOKENS)) {
-            this.completionTokens += json.get(Constants.AI.COMPLETION_TOKENS).asInt();
+            this.completionTokens += json.get(Constants.AI.COMPLETION_TOKENS).asLong();
         }
         if (json.has(Constants.AI.TOTAL_TOKENS)) {
-            this.totalTokens += json.get(Constants.AI.TOTAL_TOKENS).asInt();
+            this.totalTokens += json.get(Constants.AI.TOTAL_TOKENS).asLong();
         }
 
         JsonNode promptDetails = json.get(Constants.AI.PROMPT_TOKENS_DETAILS);
         if (promptDetails != null && promptDetails.has(Constants.AI.CACHED_TOKENS)) {
-            this.cachedTokens += promptDetails.get(Constants.AI.CACHED_TOKENS).asInt();
+            this.cachedTokens += promptDetails.get(Constants.AI.CACHED_TOKENS).asLong();
         }
 
         JsonNode completionTokensDetails = json.get(Constants.AI.COMPLETION_TOKENS_DETAILS);
         if (completionTokensDetails != null && completionTokensDetails.has(Constants.AI.REASONING_TOKENS)) {
-            this.reasoningTokens += completionTokensDetails.get(Constants.AI.REASONING_TOKENS).asInt();
+            this.reasoningTokens += completionTokensDetails.get(Constants.AI.REASONING_TOKENS).asLong();
         }
     }
 
     static AiUsage getUsageDifference(AiUsage current, AiUsage previous) {
-        int previousTotalWithoutReasoning = (previous.getCompletionTokens() - previous.getReasoningTokens() + previous.getPromptTokens());
+        long previousTotalWithoutReasoning = (previous.getCompletionTokens() - previous.getReasoningTokens() + previous.getPromptTokens());
 
         AiUsage result = new AiUsage();
         result.setPromptTokens(Math.max(current.getPromptTokens() - previousTotalWithoutReasoning, 0));
